Add arpToNmap.py file
This is a simple program written in Python, aimed for MacOS and Linux users, since some specific bash commands (lines 11 and 23) are used.

This program combines two powerful tools, "arp -a" , piping the result to "sudo nmap -O", providing a list of available hosts on the same network and also the type of Operating System they are running. At the end, it provides a simple dictionary of the MAC address and the OS associated with the specific host's IP address.

It may need some code revision and optimisation, since sometimes take too much time.

It will require Admin. permissions and it will ask for password, since we need to apply "sudo nmap -O" 

To make it executable in MacOS, once you got on the folder or directory of the file in Terminal, you just type "chmod +x arpToNmao.py" and then you should be able to execute by running "./arpToNmap.py" on command line/terminal.