Add sshConnection.py
MacOS and Linux users.

This simple program combines three tools in three different stages:

1. Performs a ping to Broadcasting IP address to stimulate all the hosts in local network. (line 12)
2. Performs an ARP to check IP and MAC addresses of all hosts inside local network. (line 15)
3. Performs an SSH connection to all of the different IP addresses gathered on step 2. The SSH connection is then closed right at the moment of connection. (lines 22-39)
4. Prints out the Host IPs, the username and password that successfully passed thru a connection.(line 33)

In order to use successfully this script, a file.txt containing a list of possible usernames and passwords must be passed as params (sys.arg[1]) on command line. The file.txt should have the following structure:

**********file.txt************
root root
admin admin
johndoe xxxxxxxxxx
janedoe password123!

***************************

To made the script executable from Terminal, once in the folder of the script it must be typed "chmod +x file.txt".

To run the script, it should work by typing the following on Terminal or bash : "./sshConnection.py file.txt"

Here's an example of the output:

192.168.1.xx johndoe xxxxxxxxxxxxx
192.168.1.xx 22 johndoe xxxxxxxxxxxxx
192.168.1.xx 22 johndoe xxxxxxxxxxxxx
SSH login successful on 192.168.1.xx:22, with username johndoe and password xxxxxxxxxxxxx

This program takes a lot of time. It needs to be improved.